UserController and quote items.
Created the User Controller. Added functions to the controller so the user is able to see all quotes assigned to them and is also able to accept new quotes that have been assigned to their account. Updated the quote model and table. I also added a quote items model and table. Pushing changes to the main branch.