Create, Edit and Deletion for products.
Implemented create, edit and deletion functionality for the products table in the project. Each function works as intended. An admin user is now able to view all the products in the database and is able to add and delete products when needed. The user is given success and error messages on each view to give clear feedback on either a deletion, create or edit request was successful.