Updated dashboard and dashboard type.
Updated the layout component to correctly show components based on if a admin user or a normal user is logged in. Placed the logic to handle this in the boot function so all views have access to this logic. Updated the dashboard to have an admin and user dashboard dependent on what user is logged in. Updated the admin controller to handle this new logic.