Project refactoring.
Refactored the layout component to handle the different admin and normal user types within the component to reduce file sizes and make code easier to navigate and update. Applied these changes to every view within the project. Added a form and button component to the project. The form is modular so it is reusable across the project as values can be added and removed when needed. Tested the project in admin and normal user state works as intended.